chg: :recycle: :truck: move make_dag inside _dag subpackage for usage in DAG.__call__
