+++ tawazi/_dag/__init__.py
@@ -1,3 +1,4 @@
+from .constructor import safe_make_dag
 from .dag import DAG, AsyncDAG, AsyncDAGExecution, DAGExecution
 
-__all__ = ["DAG", "DAGExecution", "AsyncDAG", "AsyncDAGExecution"]
+__all__ = ["DAG", "DAGExecution", "AsyncDAG", "AsyncDAGExecution", "safe_make_dag"]
