+++ tawazi/_dag/constructor.py
@@ -1,0 +1,113 @@
+from typing import Union, Callable, List, Any, Tuple, Dict
+from tawazi.node import node
+import inspect
+from tawazi.consts import RVDAG, P
+from .dag import DAG, AsyncDAG
+from tawazi.node import ArgExecNode, UsageExecNode, wrap_in_uxns, ExecNode, ReturnUXNsType
+import asyncio
+
+
+def get_args_and_default_args(func: Callable[..., Any]) -> Tuple[List[str], Dict[str, Any]]:
+    """Retrieves the arguments names and the default arguments of a function.
+
+    Args:
+        func: the target function
+
+    Returns:
+        A Tuple containing a List of argument names of non default arguments,
+         and the mapping between the arguments and their default value for default arguments
+
+    >>> def f(a1, a2, *args, d1=123, d2=None): pass
+    >>> get_args_and_default_args(f)
+    (['a1', 'a2', 'args'], {'d1': 123, 'd2': None})
+    """
+    signature = inspect.signature(func)
+    args = []
+    default_args = {}
+    for k, v in signature.parameters.items():
+        if v.default is not inspect.Parameter.empty:
+            default_args[k] = v.default
+        else:
+            args.append(k)
+
+    return args, default_args
+
+def __make_dag(_func: Callable[P, RVDAG], max_concurrency: int) -> Union[DAG[P, RVDAG], AsyncDAG[P, RVDAG]]:
+    # 2. make ExecNodes corresponding to the arguments of the ExecNode
+    # 2.1 get the names of the arguments and the default values
+    func_args, func_default_args = get_args_and_default_args(_func)
+
+    # 2.2 Construct non default arguments.
+    # Corresponding values must be provided during usage
+    args: List[ExecNode] = [
+        ArgExecNode(node.make_axn_id(arg_name, func=_func)) for arg_name in func_args
+    ]
+    # 2.2 Construct Default arguments.
+
+    for arg_name, arg in func_default_args.items():
+        axn = ArgExecNode(node.make_axn_id(arg_name, func=_func))
+        args.append(axn)
+        node.results[axn.id] = arg
+
+    # 2.3 Arguments are also ExecNodes that get executed inside the scheduler
+    node.exec_nodes.update({exec_node.id: exec_node for exec_node in args})
+    # 2.4 make UsageExecNodes for input arguments
+    uxn_args = [UsageExecNode(exec_node.id) for exec_node in args]
+
+    # 3. Execute the dependency describer function
+    # NOTE: Only ordered parameters are supported at the moment!
+    #  No **kwargs!! Only positional Arguments
+    # used to be fetch the results at the end of the computation
+    returned_val: Any = (
+        asyncio.run(_func(*uxn_args))  # type: ignore[arg-type]
+        if asyncio.iscoroutinefunction(_func)
+        else _func(*uxn_args)  # type: ignore[arg-type]
+    )
+
+    returned_usage_exec_nodes: ReturnUXNsType = wrap_in_uxns(_func, returned_val)
+
+    # 4. Construct the DAG/AsyncDAG instance
+    constructor = AsyncDAG if asyncio.iscoroutinefunction(_func) else DAG
+    return constructor(
+        results=node.results,
+        actives=node.actives,
+        exec_nodes=node.exec_nodes,
+        input_uxns=uxn_args,
+        return_uxns=returned_usage_exec_nodes,
+        max_concurrency=max_concurrency,
+    )
+
+
+def _make_dag(_func: Callable[P, RVDAG], max_concurrency: int) -> Union[DAG[P, RVDAG], AsyncDAG[P, RVDAG]]:
+    # 1. node.exec_nodes contains all the ExecNodes that concern the DAG being built at the moment.
+    #      make sure it is empty
+    node.exec_nodes = {}
+    node.results = {}
+    node.actives = {}
+    if node.DAG_PREFIX == "":
+        node.DAG_PREFIX = _func.__name__
+    else:
+        node.DAG_PREFIX += f".{_func.__name__}"
+
+    try:
+        return __make_dag(_func, max_concurrency)
+    # clean up even in case an error is raised during dag construction
+    finally:
+        # 5. Clean global variable
+        # node.exec_nodes are deep copied inside the DAG.
+        #   we can empty the global variable node.exec_nodes
+        node.exec_nodes = {}
+        node.results = {}
+        node.actives = {}
+        node.DAG_PREFIX = ""
+
+
+def safe_make_dag(
+    _func: Union[Callable[P, RVDAG]], max_concurrency: int
+) -> Union[DAG[P, RVDAG], AsyncDAG[P, RVDAG]]:
+    """Make DAG or AsyncDAG form the function that describes the DAG.
+    
+    Thread safe and cleans after itself.
+    """
+    with node.exec_nodes_lock:
+        return _make_dag(_func, max_concurrency)
