+++ tawazi/_helpers.py
@@ -1,6 +1,5 @@
 """Module for helper functions."""
-import inspect
-from typing import Any, Callable, Dict, List, NoReturn, Tuple
+from typing import Any, Callable, NoReturn
 
 import yaml
 
@@ -42,32 +41,6 @@ def ordinal(numb: int) -> str:
             else:
                 suffix = "th"
     return str(numb) + suffix
-
-
-def get_args_and_default_args(func: Callable[..., Any]) -> Tuple[List[str], Dict[str, Any]]:
-    """Retrieves the arguments names and the default arguments of a function.
-
-    Args:
-        func: the target function
-
-    Returns:
-        A Tuple containing a List of argument names of non default arguments,
-         and the mapping between the arguments and their default value for default arguments
-
-    >>> def f(a1, a2, *args, d1=123, d2=None): pass
-    >>> get_args_and_default_args(f)
-    (['a1', 'a2', 'args'], {'d1': 123, 'd2': None})
-    """
-    signature = inspect.signature(func)
-    args = []
-    default_args = {}
-    for k, v in signature.parameters.items():
-        if v.default is not inspect.Parameter.empty:
-            default_args[k] = v.default
-        else:
-            args.append(k)
-
-    return args, default_args
 
 
 def make_raise_arg_error(func_name: str, arg_name: str) -> Callable[[], NoReturn]:
